refactor: JwtService moved to services package
